+++ apps/github-search/src/pages/_app.tsx
@@ -1,11 +1,19 @@
 import React, { FC } from "react";
 import type { AppProps } from "next/app";
-import ThemeProvider from "@/common/ThemeProvider";
+import { ApolloProvider } from "@apollo/client";
+
+import ThemeProvider from "@/components/ThemeProvider";
+import Layout from "@/components/Layout";
+import client from "@/utils/apollo-client";
 
 const App: FC<AppProps> = ({ Component, pageProps }) => (
-  <ThemeProvider>
-    <Component {...pageProps} />;
-  </ThemeProvider>
+  <ApolloProvider client={client}>
+    <ThemeProvider>
+      <Layout>
+        <Component {...pageProps} />
+      </Layout>
+    </ThemeProvider>
+  </ApolloProvider>
 );
 
 export default App;